requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
Pillow
tabula-py
camelot-py[cv]
pytesseract
opencv-python-headless